requirements.txt
torch>=2.0
torchvision>=0.15
torchaudio>=2.0
transformers
Pillow
gradio

